chore: WIP stories to tsx
