--- components/button-link/css/stories/default.stories.mdx
@@ -1,30 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { ArgsTable, Canvas, Meta, Story } from "@storybook/blocks";
-import { argTypes, ButtonLink, defaultArgs, exampleArgs } from "../story-template";
-import "../index.scss";
-
-<Meta
-  id="css-button-link"
-  title="CSS Component/Link that looks like a button"
-  component={ButtonLink}
-  argTypes={argTypes}
-  args={defaultArgs}
-  parameters={{
-    status: {
-      type: "WORK IN PROGRESS",
-    },
-  }}
-/>
-
-# Button link: link die er uit ziet als button
-
-Styling met de `.utrecht-button-link` class naam:
-
-<Canvas>
-  <Story name="Default" args={exampleArgs}>
-    {ButtonLink.bind({})}
-  </Story>
-</Canvas>
-
-<ArgsTable story="Default" />
