--- components/button-link/css/stories/icon.stories.mdx
@@ -1,46 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Canvas, Meta, Story } from "@storybook/blocks";
-import { ButtonLink } from "../story-template";
-import "../index.scss";
-
-<Meta
-  id="css-button-link--icon"
-  title="CSS Component/Link that looks like a button/Icon"
-  component={ButtonLink}
-  parameters={{
-    status: {
-      type: "WORK IN PROGRESS",
-    },
-  }}
-/>
-
-# Link that looks like a button with icon
-
-## After
-
-<Canvas>
-  <Story
-    name="Icon After"
-    args={{
-      iconAfter: "utrecht-icon-arrow",
-      textContent: "Next",
-    }}
-  >
-    {ButtonLink.bind({})}
-  </Story>
-</Canvas>
-
-## Before
-
-<Canvas>
-  <Story
-    name="Icon Before"
-    args={{
-      iconBefore: "utrecht-icon-checkmark",
-      textContent: "Versturen",
-    }}
-  >
-    {ButtonLink.bind({})}
-  </Story>
-</Canvas>
