--- components/button-link/css/stories/readme.stories.mdx
@@ -1,8 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Markdown, Meta } from "@storybook/blocks";
-import document from "../../README.md";
-
-<Meta id="css-button-link--readme" title="CSS Component/Link that looks like a button/README" />
-
-<Markdown>{document}</Markdown>
