--- components/checkbox/css/stories/default.stories.mdx
@@ -1,25 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { ArgsTable, Canvas, Meta, Story } from "@storybook/blocks";
-import { argTypes, Checkbox, defaultArgs } from "../story-template";
-
-<Meta
-  id="css-checkbox"
-  title="CSS Component/Checkbox"
-  component={Checkbox}
-  argTypes={argTypes}
-  args={defaultArgs}
-  parameters={{
-    status: {
-      type: "WORK IN PROGRESS",
-    },
-  }}
-/>
-
-# Checkbox Component
-
-<Canvas>
-  <Story name="Default">{Checkbox.bind({})}</Story>
-</Canvas>
-
-<ArgsTable story="Default" />
