--- components/checkbox/css/stories/readme.stories.mdx
@@ -1,8 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Markdown, Meta } from "@storybook/blocks";
-import document from "../../README.md";
-
-<Meta id="css-checkbox--readme" title="CSS Component/Checkbox/README" />
-
-<Markdown>{document}</Markdown>
