--- components/checkbox/css/stories/state.stories.mdx
@@ -1,57 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Canvas, Meta, Story } from "@storybook/blocks";
-import { Checkbox } from "../story-template";
-
-<Meta
-  id="css-checkbox--state"
-  title="CSS Component/Checkbox/States"
-  component={Checkbox}
-  parameters={{
-    status: {
-      type: "WORK IN PROGRESS",
-    },
-  }}
-/>
-
-# Checkbox Component States
-
-## Checked
-
-<Canvas>
-  <Story
-    name="Checked"
-    args={{
-      checked: true,
-    }}
-  >
-    {Checkbox.bind({})}
-  </Story>
-</Canvas>
-
-## Disabled
-
-<Canvas>
-  <Story
-    name="Disabled"
-    args={{
-      disabled: true,
-    }}
-  >
-    {Checkbox.bind({})}
-  </Story>
-</Canvas>
-
-## Checked and Disabled
-
-<Canvas>
-  <Story
-    name="Checked and Disabled"
-    args={{
-      checked: true,
-      disabled: true,
-    }}
-  >
-    {Checkbox.bind({})}
-  </Story>
-</Canvas>
