--- components/checkbox/css/stories/tokens.stories.mdx
@@ -1,16 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Meta } from "@storybook/blocks";
-import { ComponentTokensSection } from "../../../../documentation/components/ComponentTokensSection";
-import tokens from "../../../../proprietary/design-tokens/dist/index.json";
-import tokensDefinition from "../../tokens.json";
-
-<Meta id="css-checkbox--tokens" title="CSS Component/Checkbox/Design Tokens" />
-
-# Checkbox Design Tokens
-
-<ComponentTokensSection
-  tokens={tokens}
-  definition={tokensDefinition}
-  component="utrecht-checkbox"
-></ComponentTokensSection>
