--- components/code/css/stories/default.stories.mdx
@@ -1,30 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { ArgsTable, Canvas, Meta, Story } from "@storybook/blocks";
-import { argTypes, Code, defaultArgs, exampleArgs } from "../story-template";
-import "../index.scss";
-
-<Meta
-  id="css-code"
-  title="CSS Component/Code"
-  component={Code}
-  argTypes={argTypes}
-  args={defaultArgs}
-  parameters={{
-    status: {
-      type: "WORK IN PROGRESS",
-    },
-  }}
-/>
-
-# Code
-
-Styling via the `.utrecht-code` class name:
-
-<Canvas>
-  <Story name="Default" args={exampleArgs}>
-    {Code.bind({})}
-  </Story>
-</Canvas>
-
-<ArgsTable story="Default" />
