--- components/code/css/stories/link.stories.mdx
@@ -1,36 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Canvas, Meta, Story } from "@storybook/blocks";
-import { CodeInLink } from "../story-template";
-
-<Meta
-  id="css-code--link"
-  title="CSS Component/Code"
-  component={CodeInLink}
-  parameters={{
-    status: {
-      type: "WORK IN PROGRESS",
-    },
-    docs: {
-      source: {
-        state: "closed",
-      },
-    },
-  }}
-/>
-
-# Code in link
-
-De code die een link is moet herkenbaar zijn als link.
-
-<Canvas>
-  <Story
-    name="Code in link"
-    args={{
-      href: "https://w3c.github.io/csswg-drafts/mediaqueries-5/#prefers-color-scheme",
-      textContent: "@media (prefers-color-scheme: dark) {}",
-    }}
-  >
-    {CodeInLink.bind({})}
-  </Story>
-</Canvas>
