--- components/code/css/stories/paragraph.stories.mdx
@@ -1,35 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Canvas, Meta, Story } from "@storybook/blocks";
-import { argTypes, CodeInParagraph, defaultArgs } from "../story-template";
-
-<Meta
-  id="css-code--paragraph"
-  title="CSS Component/Code"
-  component={CodeInParagraph}
-  argTypes={argTypes}
-  args={defaultArgs}
-  parameters={{
-    status: {
-      type: "WORK IN PROGRESS",
-    },
-    docs: {
-      source: {
-        state: "closed",
-      },
-    },
-  }}
-/>
-
-# Code in a paragraph
-
-<Canvas>
-  <Story
-    name="Code in a paragraph"
-    args={{
-      textContent: '<input type="url" value="https://example.fi/">',
-    }}
-  >
-    {CodeInParagraph()}
-  </Story>
-</Canvas>
