--- components/code/css/stories/tokens.stories.mdx
@@ -1,12 +1,0 @@
-{/* @license CC0-1.0 */}
-
-import { Meta } from "@storybook/blocks";
-import { ComponentTokensSection } from "../../../../documentation/components/ComponentTokensSection";
-import tokens from "../../../../proprietary/design-tokens/dist/index.json";
-import tokensDefinition from "../../tokens.json";
-
-<Meta id="css-code--tokens" title="CSS Component/Code/Design Tokens" />
-
-# Code Design Tokens
-
-<ComponentTokensSection tokens={tokens} definition={tokensDefinition} component="utrecht-code"></ComponentTokensSection>
